完成问卷，问题及用户管理模块

+++ OnlineSurvey/Quick.Common/Models/PageRequest.cs
@@ -38,9 +38,9 @@ namespace Quick.Common.Models
         /// </summary>
         public int PageIndex { get; set; } = 1;
         /// <summary>
-        /// 总页数
+        /// 总页数 Math.Ceiling(Convert.ToDecimal(TotalCount) / PageSize);
         /// </summary>
-        public int TotalPage { get; set; }
+        public decimal TotalPage { get; set; } 
         /// <summary>
         /// 总记录数
         /// </summary>
